+++ packages/telescope-lib/lib/templates.js
@@ -1,4 +1,4 @@
-Telescope.config.customPrefixes = ["custom_"];
+Telescope.config.customPrefixes = [];
 
 Telescope.config.addCustomPrefix = function (prefix) {
   Telescope.config.customPrefixes.push(prefix);